split quantization API to dynamic and static specifically
