+++ onnxruntime/python/tools/quantization/__init__.py
@@ -1,4 +1,5 @@
-from .quantize import quantize
+from .quantize import quantize, quantize_static, quantize_dynamic
 from .quantize import QuantizationMode
 from .calibrate import CalibrationDataReader
 from .calibrate import calibrate
+from .quant_utils import QuantType
